requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
cloudinary
gunicorn
flask run
psycopg2-binary
cloudinary
email-validator
flask
flask-login
flask-sqlalchemy
gunicorn
psycopg2-binary
python-dotenv
requests
sqlalchemy
werkzeug